[refactor] bottom filter 필터 적용 & 리셋 기능 refactoring
