+++ src/components/Common/Filter/BottomFilter.tsx
@@ -1,14 +1,86 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import StyleFilter from './StyleFilter';
-import GenderFilter from './GenderFilter';
-import AgeFilter from './AgeFilter';
-import ColorFilter from './ColorFilter';
-import RangeFilter from './RangeFilter';
-import { bottomSheetTabs } from 'constant';
 import { IBottomFilterProps } from 'types/Common';
-import { useAppDispatch } from 'redux/store';
-import { resetFilters, setFilters } from 'redux/reducer/filter';
+import useSetFilters from 'hooks/useSetFilters';
+import { bottomSheetTabs } from 'constant/bottomFilters';
+import { tabItems } from 'constant/constantComponents/bottomFilters';
+
+function BottomFilter({ setIsFilterOpen }: IBottomFilterProps) {
+  const [activeTab, setActiveTab] = useState(0);
+  const onClickTab = (e: React.MouseEvent<HTMLLIElement>, tabId: number) => {
+    setActiveTab(tabId);
+  };
+
+  const { setFilter, resetFilter } = useSetFilters();
+
+  const onClickCloseFilter = () => {
+    setIsFilterOpen(false);
+  };
+
+  const onClickBottomSheetMain = (e: React.MouseEvent<HTMLDivElement>) => {
+    e.stopPropagation();
+  };
+
+  const onClickFilterReset = (
+    e: React.MouseEvent<HTMLButtonElement>,
+    activeTab: number,
+  ) => {
+    e.preventDefault();
+    resetFilter(activeTab);
+  };
+
+  const onClickSetFilter = (
+    e: React.MouseEvent<HTMLButtonElement>,
+    activeTab: number,
+  ) => {
+    e.preventDefault();
+    setFilter(activeTab);
+    setIsFilterOpen(false);
+  };
+
+  return (
+    <BottomSheet onClick={onClickCloseFilter}>
+      <BottomSheetMain onClick={onClickBottomSheetMain}>
+        <BottomSheetCategory>
+          <BottomSheetCategorys>
+            {bottomSheetTabs.map((tab) => (
+              <TabCategory
+                $active={tab.id === activeTab}
+                key={tab.id}
+                onClick={(e) => onClickTab(e, tab.id)}
+              >
+                {tab.name}
+              </TabCategory>
+            ))}
+          </BottomSheetCategorys>
+        </BottomSheetCategory>
+        <form>
+          <div>{tabItems[activeTab].content}</div>
+          <BottomSheetBtns>
+            <ul>
+              <li>
+                <BottomSheetReset
+                  onClick={(e) => onClickFilterReset(e, activeTab)}
+                >
+                  {tabItems[activeTab].name} 재설정
+                </BottomSheetReset>
+              </li>
+              <li>
+                <BottomSheetSubmit
+                  onClick={(e) => onClickSetFilter(e, activeTab)}
+                >
+                  옷장 구경하기
+                </BottomSheetSubmit>
+              </li>
+            </ul>
+          </BottomSheetBtns>
+        </form>
+      </BottomSheetMain>
+    </BottomSheet>
+  );
+}
+
+export default BottomFilter;
 
 const BottomSheet = styled.div`
   position: fixed;
@@ -108,96 +180,3 @@ const BottomSheetSubmit = styled.button`
   background-color: ${({ theme }) => theme.colors.white};
   border-radius: ${({ theme }) => theme.borderRadius.borderRadius50};
 `;
-
-function BottomFilter({ setIsFilterOpen }: IBottomFilterProps) {
-  const [activeTab, setActiveTab] = useState(0);
-  const onClickTab = (e: React.MouseEvent<HTMLLIElement>, tabId: number) => {
-    setActiveTab(tabId);
-  };
-  const dispatch = useAppDispatch();
-
-  const tabItems = [
-    { name: '성별', id: 0, content: <GenderFilter filterIndex={0} /> },
-    { name: '스타일', id: 1, content: <StyleFilter filterIndex={1} /> },
-    { name: '연령대', id: 2, content: <AgeFilter filterIndex={2} /> },
-    { name: '컬러', id: 3, content: <ColorFilter filterIndex={3} /> },
-    {
-      name: '키',
-      id: 4,
-      content: <RangeFilter filterType="height" filterIndex={4} />,
-    },
-    {
-      name: '몸무게',
-      id: 5,
-      content: <RangeFilter filterType="weight" filterIndex={5} />,
-    },
-  ];
-
-  const onClickCloseFilter = () => {
-    setIsFilterOpen(false);
-  };
-
-  const onClickBottomSheetMain = (e: React.MouseEvent<HTMLDivElement>) => {
-    e.stopPropagation();
-  };
-
-  const onClickFilterReset = (
-    e: React.MouseEvent<HTMLButtonElement>,
-    activeTab: number,
-  ) => {
-    e.preventDefault();
-    dispatch(resetFilters(activeTab));
-  };
-
-  const onClickSetFilter = (
-    e: React.MouseEvent<HTMLButtonElement>,
-    activeTab: number,
-  ) => {
-    e.preventDefault();
-    dispatch(setFilters(activeTab));
-    setIsFilterOpen(false);
-  };
-
-  return (
-    <BottomSheet onClick={onClickCloseFilter}>
-      <BottomSheetMain onClick={onClickBottomSheetMain}>
-        <BottomSheetCategory>
-          <BottomSheetCategorys>
-            {bottomSheetTabs.map((tab) => (
-              <TabCategory
-                $active={tab.id === activeTab}
-                key={tab.id}
-                onClick={(e) => onClickTab(e, tab.id)}
-              >
-                {tab.name}
-              </TabCategory>
-            ))}
-          </BottomSheetCategorys>
-        </BottomSheetCategory>
-        <form>
-          <div>{tabItems[activeTab].content}</div>
-          <BottomSheetBtns>
-            <ul>
-              <li>
-                <BottomSheetReset
-                  onClick={(e) => onClickFilterReset(e, activeTab)}
-                >
-                  {tabItems[activeTab].name} 재설정
-                </BottomSheetReset>
-              </li>
-              <li>
-                <BottomSheetSubmit
-                  onClick={(e) => onClickSetFilter(e, activeTab)}
-                >
-                  옷장 구경하기
-                </BottomSheetSubmit>
-              </li>
-            </ul>
-          </BottomSheetBtns>
-        </form>
-      </BottomSheetMain>
-    </BottomSheet>
-  );
-}
-
-export default BottomFilter;
