+++ src/constant/bottomFilters.ts
@@ -1,0 +1,67 @@
+export const activityData = [
+  {
+    name: 'VACATION',
+    type: 'vacation',
+  },
+  {
+    name: '수상레져',
+    type: 'leisure',
+  },
+  {
+    name: '골프',
+    type: 'golf',
+  },
+  {
+    name: '등산',
+    type: 'climbing',
+  },
+];
+
+export const styleCategory = [
+  { name: '캐주얼', id: 0 },
+  { name: '모던', id: 1 },
+  { name: '스트릿', id: 2 },
+  { name: '시크', id: 3 },
+  { name: '로맨틱', id: 4 },
+  { name: '걸리시', id: 5 },
+  { name: '레트로', id: 6 },
+  { name: '댄디', id: 7 },
+  { name: '스포츠', id: 8 },
+  { name: '아메리칸 캐주얼', id: 9 },
+  { name: '포멀', id: 10 },
+];
+
+export const colorCategory = [
+  { name: 'black', id: 0 },
+  { name: 'gray', id: 1 },
+  { name: 'white', id: 2 },
+  { name: 'gold', id: 3 },
+  { name: 'silver', id: 4 },
+  { name: 'red', id: 5 },
+  { name: 'blue', id: 6 },
+  { name: 'green', id: 7 },
+  { name: 'navy', id: 8 },
+  { name: 'brown', id: 9 },
+  { name: 'beige', id: 10 },
+  { name: 'yellow', id: 11 },
+  { name: 'orange', id: 12 },
+  { name: 'pink', id: 13 },
+  { name: 'purple', id: 14 },
+];
+
+export const bottomSheetTabs = [
+  { name: '성별', id: 0 },
+  { name: '스타일', id: 1 },
+  { name: '연령대', id: 2 },
+  { name: '컬러', id: 3 },
+  { name: '키', id: 4 },
+  { name: '몸무게', id: 5 },
+];
+
+export const ageCategory = [
+  { value: 10, id: 0 },
+  { value: 20, id: 1 },
+  { value: 30, id: 2 },
+];
+
+export const RANGE_GAP = 10;
